mpesa::services

Struct B2cBuilder

pub struct B2cBuilder<'mpesa> { /* private fields */ }

B2C transaction builder struct

Implementations

impl<'mpesa> B2cBuilder<'mpesa>

pub fn new( client: &'mpesa Mpesa, initiator_name: &'mpesa str, ) -> B2cBuilder<'mpesa>

Create a new B2C builder. Requires an initiator_name, the credential/ username used to authenticate the transaction request

pub fn originator_conversation_id( self, originator_conversation_id: &'mpesa str, ) -> B2cBuilder<'mpesa>

Adds the OriginatorConversationID.

pub fn command_id(self, command_id: CommandId) -> B2cBuilder<'mpesa>

Adds the CommandId. Defaults to CommandId::BusinessPayment if not explicitly provided.

pub fn party_a(self, party_a: &'mpesa str) -> B2cBuilder<'mpesa>

Adds Party A which is a required field Party A should be a paybill number.

Errors

If Party A is invalid or not provided

pub fn party_b(self, party_b: &'mpesa str) -> B2cBuilder<'mpesa>

Adds Party B which is a required field Party B should be a mobile number.

Errors

If Party B is invalid or not provided

pub fn parties( self, party_a: &'mpesa str, party_b: &'mpesa str, ) -> B2cBuilder<'mpesa>

๐Deprecated

Adds Party A and Party B. Both are required fields Party A should be a paybill number while Party B should be a mobile number.

Errors

If either Party A or Party B is invalid or not provided

pub fn remarks(self, remarks: &'mpesa str) -> B2cBuilder<'mpesa>

Adds Remarks. This is an optional field, will default to โNoneโ if not explicitly provided

pub fn occasion(self, occasion: &'mpesa str) -> B2cBuilder<'mpesa>

Adds Occasion. This is an optional field, will default to an empty string

pub fn amount<Number: Into<f64>>(self, amount: Number) -> B2cBuilder<'mpesa>

Adds an amount to the request This is a required field

pub fn timeout_url(self, timeout_url: &'mpesa str) -> B2cBuilder<'mpesa>

Error

If QueueTimeoutUrl is invalid or not provided

pub fn result_url(self, result_url: &'mpesa str) -> B2cBuilder<'mpesa>

Error

If ResultUrl is invalid or not provided

pub fn urls( self, timeout_url: &'mpesa str, result_url: &'mpesa str, ) -> B2cBuilder<'mpesa>

๐Deprecated

Adds QueueTimeoutUrl and ResultUrl. This is a required field

Error

If either QueueTimeoutUrl and ResultUrl is invalid or not provided

pub async fn send(self) -> MpesaResult<B2cResponse>

B2C API

Sends b2c payment request.

This API enables Business to Customer (B2C) transactions between a company and customers who are the end-users of its products or services. Use of this API requires a valid and verified B2C M-Pesa Short code. See more here

A successful request returns a B2cResponse type

Errors

Returns a MpesaError on failure.