mpesa::services

Struct MpesaExpressBuilder

pub struct MpesaExpressBuilder<'mpesa> { /* private fields */ }

Builder for MpesaExpress.

Implementations

impl<'mpesa> MpesaExpressBuilder<'mpesa>

pub fn client(&self, value: &'mpesa Mpesa) -> Self

pub fn business_short_code<VALUE: Into<&'mpesa str>>( &mut self, value: VALUE, ) -> &mut Self

This is the organizationโs shortcode (Paybill or Buygoods - A 5 to 6-digit account number) used to identify an organization and receive the transaction.

pub fn transaction_type(&mut self, value: CommandId) -> &mut Self

This is the transaction type that is used to identify the transaction when sending the request to M-PESA

The TransactionType for Mpesa Express is either CommandId::BusinessBuyGoods or CommandId::CustomerPayBillOnline or CommandId::CustomerBuyGoodsOnline

pub fn amount(&mut self, value: u32) -> &mut Self

This is the Amount transacted normally a numeric value

pub fn party_a(&mut self, value: &'mpesa str) -> &mut Self

The phone number sending money.

pub fn party_b(&mut self, value: &'mpesa str) -> &mut Self

The organization that receives the funds This should be the Till Number in the case of CommandId::CustomerBuyGoodsOnline

pub fn phone_number(&mut self, value: &'mpesa str) -> &mut Self

The Mobile Number to receive the STK Pin Prompt.

pub fn callback_url<VALUE: Into<Url>>(&mut self, value: VALUE) -> &mut Self

A CallBack URL is a valid secure URL that is used to receive notifications from M-Pesa API. It is the endpoint to which the results will be sent by M-Pesa API.

pub fn try_callback_url<VALUE: TryInto<Url>>( &mut self, value: VALUE, ) -> Result<&mut Self, VALUE::Error>

A CallBack URL is a valid secure URL that is used to receive notifications from M-Pesa API. It is the endpoint to which the results will be sent by M-Pesa API.

pub fn account_ref<VALUE: Into<&'mpesa str>>( &mut self, value: VALUE, ) -> &mut Self

Account Reference: This is an Alpha-Numeric parameter that is defined by your system as an Identifier of the transaction for CustomerPayBillOnline

pub fn transaction_desc<VALUE: Into<&'mpesa str>>( &mut self, value: VALUE, ) -> &mut Self

This is any additional information/comment that can be sent along with the request from your system

pub fn pass_key<VALUE: Into<&'mpesa str>>(&mut self, value: VALUE) -> &mut Self

This is the password used for encrypting the request sent: The password for encrypting the request is obtained by base64 encoding BusinessShortCode, Passkey and Timestamp. The timestamp format is YYYYMMDDHHmmss

pub fn build(&self) -> Result<MpesaExpress<'mpesa>, MpesaError>

Builds a new MpesaExpress.

Errors

If a required field has not been initialized.