Struct mpesa::services::TransactionStatusBuilder

pub struct TransactionStatusBuilder<'mpesa, Env: ApiEnvironment> { /* private fields */ }

Implementations

impl<'mpesa, Env: ApiEnvironment> TransactionStatusBuilder<'mpesa, Env>

pub fn new( client: &'mpesa Mpesa<Env>, initiator: &'mpesa str ) -> TransactionStatusBuilder<'mpesa, Env>

Creates new TransactionStatusBuilder

pub fn command_id(self, command_id: CommandId) -> Self

Adds CommandId. Defaults to CommandId::TransactionStatus if no value explicitly passed

Errors

If CommandId is not valid

pub fn transaction_id(self, transaction_id: &'mpesa str) -> Self

Add the Mpesa Transaction ID of the transaction which you wish to reverse

This is a required field.

pub fn party_a(self, party_a: &'mpesa str) -> Self

Organization receiving the transaction

This is required field

pub fn identifier_type(self, identifier_type: IdentifierTypes) -> Self

Type of organization receiving the transaction

This is an optional field, defaults to IdentifierTypes::ShortCode

pub fn result_url(self, result_url: &'mpesa str) -> Self

Error

If ResultUrl is invalid or not provided

pub fn timeout_url(self, timeout_url: &'mpesa str) -> Self

Adds QueueTimeoutUrl and ResultUrl. This is a required field

Error

If either QueueTimeoutUrl and ResultUrl is invalid or not provided

pub fn remarks(self, remarks: &'mpesa str) -> Self

Comments that are sent along with the transaction.

This is an optional field, defaults to “None”

pub fn occasion(self, occasion: &'mpesa str) -> Self

Adds any additional information to be associated with the transaction.

This is an optional Parameter, defaults to “None”

pub async fn send(self) -> MpesaResult<TransactionStatusResponse>

Transaction Status API

Requests for the status of a transaction

This API enables the status of a B2B, B2C or C2B M-Pesa transaction Required parameters:

transaction_id: This is the Mpesa Transaction ID of the transaction which you wish to reverse

See more from the Safaricom API docs here

A successful request returns a TransactionStatusResponse type

Errors

Returns a MpesaError on failure.