mpesa

Module services

MPESA Services

Using the builder pattern in this to procedurally build Mpesa service specific payloads which are ultimately consumed and the request sent by calling the send method. Some of the builder methods for certain services are optional with default values standing in their place when the builder gets consumed

Here are the currently supported services:

Account Balance

B2B

B2C

C2B Register

C2B Simulate

Mpesa Express/ STK Push

Transaction Reversal

Bill Manager

Transaction Status

Dynamic QR

Structs

AccountBalanceBuilder

AccountBalanceResponse

B2bBuilder

B2B transaction builder struct

B2bResponse

B2cBuilder

B2C transaction builder struct

B2cResponse

BulkInvoiceBuilder

BulkInvoiceResponse

C2bRegisterBuilder

C2B Register builder

C2bRegisterResponse

C2bSimulateBuilder

C2bSimulateResponse

CancelInvoiceBuilder

CancelInvoiceResponse

DynamicQR

Dynamic QR builder struct

DynamicQRBuilder

Builder for DynamicQR.

DynamicQRRequest

DynamicQRResponse

MpesaExpress

MpesaExpressBuilder

Builder for MpesaExpress.

MpesaExpressQuery

MpesaExpressQueryBuilder

Builder for MpesaExpressQuery.

MpesaExpressQueryResponse

MpesaExpressRequest

MpesaExpressResponse

OnboardBuilder

OnboardModifyBuilder

OnboardModifyResponse

OnboardResponse

ReconciliationBuilder

ReconciliationResponse

SingleInvoiceBuilder

SingleInvoiceResponse

TransactionReversal

TransactionReversalBuilder

Builder for TransactionReversal.

TransactionReversalRequest

TransactionReversalResponse

TransactionStatusBuilder

TransactionStatusResponse