mpesa

Struct Mpesa

pub struct Mpesa { /* private fields */ }

Mpesa client that will facilitate communication with the Safaricom API

Implementations

impl Mpesa

pub fn new<S: Into<String>>( consumer_key: S, consumer_secret: S, environment: impl ApiEnvironment, ) -> Self

Constructs a new Mpesa client.

Example

use mpesa::{Environment, Mpesa};

#[tokio::main]
async fn main() {
    dotenvy::dotenv().ok();

    let client = Mpesa::new(
        dotenvy::var("CONSUMER_KEY").unwrap(),
        dotenvy::var("CONSUMER_SECRET").unwrap(),
        Environment::Sandbox,
    );

    assert!(client.is_connected().await);
}

Panics

This method can panic if a TLS backend cannot be initialized for the internal http_client

pub fn set_initiator_password<S: Into<String>>(&self, initiator_password: S)

Optional in development but required for production for the following apis:

account_balance

b2b

b2c

transaction_reversal

transaction_status

You will need to call this method and set your production initiator password. If in development, a default initiator password from the test credentials is already pre-set

Example

use mpesa::{Environment, Mpesa};

#[tokio::main]
async fn main() {
    dotenvy::dotenv().ok();

    let client = Mpesa::new(
        dotenvy::var("CONSUMER_KEY").unwrap(),
        dotenvy::var("CONSUMER_SECRET").unwrap(),
        Environment::Sandbox,
    );
    client.set_initiator_password("your_initiator_password");
    assert!(client.is_connected().await);
}

pub fn has_cached_auth(&self) -> bool

Check if we have a cached valid auth token

pub async fn is_connected(&self) -> bool

Checks if the client can be authenticated

pub fn b2c<'a>(&'a self, initiator_name: &'a str) -> B2cBuilder<'a>

B2C

Requires an initiator_name, the credential/ username used to authenticate the transaction request Returns a B2cBuilder for building a B2C transaction struct.

Safaricom the API docs B2C reference.

Example

use mpesa::{Environment, Mpesa};

#[tokio::main]
async fn main() {
    dotenvy::dotenv().ok();

    let client = Mpesa::new(
        dotenvy::var("CONSUMER_KEY").unwrap(),
        dotenvy::var("CONSUMER_SECRET").unwrap(),
        Environment::Sandbox,
    );

    let response = client
        .b2c("testapi496")
        .originator_conversation_id("77a98672-7782-4e04-9be2-d6880fb3e6a5")
        .party_a("600496")
        .party_b("254708374149")
        .result_url("https://testdomain.com/err")
        .timeout_url("https://testdomain.com/ok")
        .amount(1000)
        .remarks("Your Remark") // optional, defaults to "None"
        .occasion("Your Occasion") // optional, defaults to "None"
        .command_id(mpesa::CommandId::BusinessPayment) // optional, defaults to `CommandId::BusinessPayment`
        .send()
        .await;
    assert!(response.is_ok())
}

pub fn b2b<'a>(&'a self, initiator_name: &'a str) -> B2bBuilder<'a>

B2B

This API enables you to pay bills directly from your business account to a pay bill number, or a paybill store. You can use this API to pay on behalf of a consumer/requester.

The transaction moves money from your MMF/Working account to the recipient’s utility account.

Returns a B2bBuilder. Requires an initiator_name, the credential/ username used to authenticate the transaction request

Safaricom API docs B2B reference

Example

use mpesa::{Environment, Mpesa};

#[tokio::main]
async fn main() {
    dotenvy::dotenv().ok();

    let client = Mpesa::new(
        dotenvy::var("CONSUMER_KEY").unwrap(),
        dotenvy::var("CONSUMER_SECRET").unwrap(),
        Environment::Sandbox,
    );

    let response = client
        .b2b("testapi496")
        .party_a("600496")
        .party_b("600000")
        .result_url("https://testdomain.com/err")
        .timeout_url("https://testdomain.com/ok")
        .account_ref("254708374149")
        .amount(1000)
        .command_id(mpesa::CommandId::BusinessToBusinessTransfer) // optional, defaults to `CommandId::BusinessToBusinessTransfer`
        .remarks("None") // optional, defaults to "None"
        .sender_id(mpesa::IdentifierTypes::ShortCode) // optional, defaults to `IdentifierTypes::ShortCode`
        .receiver_id(mpesa::IdentifierTypes::ShortCode) // optional, defaults to `IdentifierTypes::ShortCode`
        .send()
        .await;

    assert!(response.is_ok());
}

pub fn onboard(&self) -> OnboardBuilder<'_>

Onboard

Creates a OnboardBuilder which allows you to opt in as a biller to the bill manager features.

Safaricom API docs reference

Example

use mpesa::{Mpesa, Environment, SendRemindersTypes};

#[tokio::main]
async fn main() {
    dotenvy::dotenv().ok();

    let client = Mpesa::new(
        dotenvy::var("CONSUMER_KEY").unwrap(),
        dotenvy::var("CONSUMER_SECRET").unwrap(),
        Environment::Sandbox,
    );

    let response = client
        .onboard()
        .callback_url("https://testdomain.com/true")
        .email("email@test.com")
        .logo("https://file.domain/file.png")
        .official_contact("0712345678")
        .send_reminders(SendRemindersTypes::Enable)
        .short_code("718003")
        .send()
        .await;

    assert!(response.is_ok());
}

pub fn onboard_modify(&self) -> OnboardModifyBuilder<'_>

Onboard Modify

Creates a OnboardModifyBuilder which allows you to opt in as a biller to the bill manager features.

Safaricom API docs reference

Example

use mpesa::{Mpesa, Environment, SendRemindersTypes};

#[tokio::main]
async fn main() {
    dotenvy::dotenv().ok();

    let client = Mpesa::new(
        dotenvy::var("CONSUMER_KEY").unwrap(),
        dotenvy::var("CONSUMER_SECRET").unwrap(),
        Environment::Sandbox,
    );

    let response = client
        .onboard_modify()
        .callback_url("https://testdomain.com/true")
        .email("email@test.com")
        .logo("https://file.domain/file.png")
        .official_contact("0712345678")
        .send_reminders(SendRemindersTypes::Enable)
        .short_code("600496")
        .send()
        .await;

    assert!(response.is_ok());
}

pub fn bulk_invoice(&self) -> BulkInvoiceBuilder<'_>

Bulk Invoice

Creates a BulkInvoiceBuilder which allows you to send invoices to your customers in bulk.

Safaricom API docs reference

Example

use mpesa::{Mpesa, Environment, Invoice, InvoiceItem};
use chrono::prelude::Utc;

#[tokio::main]
async fn main() {
    dotenvy::dotenv().ok();

    let client = Mpesa::new(
        dotenvy::var("CONSUMER_KEY").unwrap(),
        dotenvy::var("CONSUMER_SECRET").unwrap(),
        Environment::Sandbox,
    );

    let response = client
        .bulk_invoice()
        // Add multiple invoices at once
        .invoices(vec![
            Invoice {
                amount: 1000.0,
                account_reference: "John Doe",
                billed_full_name: "John Doe",
                billed_period: "August 2021",
                billed_phone_number: "0712345678",
                due_date: Utc::now(),
                external_reference: "INV2345",
                invoice_items: Some(
                    vec![InvoiceItem {amount: 1000.0, item_name: "An item"}]
                ),
                invoice_name: "Invoice 001"
            }
        ])
        // Add a single invoice
        .invoice(
            Invoice {
                amount: 1000.0,
                account_reference: "John Doe",
                billed_full_name: "John Doe",
                billed_period: "August 2021",
                billed_phone_number: "0712345678",
                due_date: Utc::now(),
                external_reference: "INV2345",
                invoice_items: Some(vec![InvoiceItem {
                    amount: 1000.0,
                    item_name: "An item",
                }]),
                invoice_name: "Invoice 001",
            }
        )
        .send()
        .await;

    assert!(response.is_ok());
}

pub fn single_invoice(&self) -> SingleInvoiceBuilder<'_>

Single Invoice

Creates a SingleInvoiceBuilder which allows you to create and send invoices to your customers.

Safaricom API docs reference

Example

use chrono::prelude::Utc;
use mpesa::{Environment, InvoiceItem, Mpesa};

#[tokio::main]
async fn main() {
    dotenvy::dotenv().ok();

    let client = Mpesa::new(
        dotenvy::var("CONSUMER_KEY").unwrap(),
        dotenvy::var("CONSUMER_SECRET").unwrap(),
        Environment::Sandbox,
    );

    let response = client
        .single_invoice()
        .amount(1000.0)
        .account_reference("John Doe")
        .billed_full_name("John Doe")
        .billed_period("August 2021")
        .billed_phone_number("0712345678")
        .due_date(Utc::now())
        .external_reference("INV2345")
        .invoice_items(vec![InvoiceItem {
            amount: 1000.0,
            item_name: "An item",
        }])
        .invoice_name("Invoice 001")
        .send()
        .await;

    assert!(response.is_ok());
}

pub fn reconciliation(&self) -> ReconciliationBuilder<'_>

Reconciliation

Creates a ReconciliationBuilder which enables your customers to receive e-receipts for payments made to your paybill account.

Safaricom API docs reference

Example

use mpesa::{Mpesa, Environment};
use chrono::prelude::Utc;

#[tokio::main]
async fn main() {
    dotenvy::dotenv().ok();

    let client = Mpesa::new(
        dotenvy::var("CONSUMER_KEY").unwrap(),
        dotenvy::var("CONSUMER_SECRET").unwrap(),
        Environment::Sandbox,
    );

    let response = client
        .reconciliation()
        .account_reference("John Doe")
        .external_reference("INV2345")
        .full_name("John Doe")
        .invoice_name("Invoice 001")
        .paid_amount(1000.0)
        .payment_date(Utc::now())
        .phone_number("0712345678")
        .transaction_id("TRANSACTION_ID")
        .send()
        .await;

    assert!(response.is_ok());
}

pub fn cancel_invoice(&self) -> CancelInvoiceBuilder<'_>

Cancel Invoice

Creates a CancelInvoiceBuilder which allows you to recall a sent invoice.

Safaricom API docs reference

Example

use mpesa::{Mpesa, Environment, SendRemindersTypes};
use chrono::prelude::Utc;

#[tokio::main]
async fn main() {
    dotenvy::dotenv().ok();

    let client = Mpesa::new(
        dotenvy::var("CONSUMER_KEY").unwrap(),
        dotenvy::var("CONSUMER_SECRET").unwrap(),
        Environment::Sandbox,
    );

    let response = client
        .cancel_invoice()
        .external_references(vec!["9KLSS011"])
        .send()
        .await;

    assert!(response.is_ok());
}

pub fn c2b_register(&self) -> C2bRegisterBuilder<'_>

C2B Register

Register URL API works hand in hand with Customer to Business (C2B) APIs and allows receiving payment notifications to your paybill. This API enables you to register the callback URLs via which you shall receive notifications for payments to your pay bill/till number.

There are two URLs required for Register URL API: Validation URL and Confirmation URL.

Returns a C2bRegisterBuilder

See more from the Safaricom API docs C2B Register URL

Example

use mpesa::{Environment, Mpesa};

#[tokio::main]
async fn main() {
    dotenvy::dotenv().ok();

    let client = Mpesa::new(
        dotenvy::var("CONSUMER_KEY").unwrap(),
        dotenvy::var("CONSUMER_SECRET").unwrap(),
        Environment::Sandbox,
    );

    let response = client
        .c2b_register()
        .short_code("600496")
        .confirmation_url("https://testdomain.com/true")
        .validation_url("https://testdomain.com/valid")
        .response_type(mpesa::ResponseType::Completed) // optional, defaults to `ResponseTypes::Complete`
        .send()
        .await;

    assert!(response.is_ok())
}

pub fn c2b_simulate(&self) -> C2bSimulateBuilder<'_>

C2B Simulate

Creates a C2bSimulateBuilder for simulating C2B transactions

See more Simulate C2B

Example

use mpesa::{Environment, Mpesa};

#[tokio::main]
async fn main() {
    dotenvy::dotenv().ok();

    let client = Mpesa::new(
        dotenvy::var("CONSUMER_KEY").unwrap(),
        dotenvy::var("CONSUMER_SECRET").unwrap(),
        Environment::Sandbox,
    );

    let response = client
        .c2b_simulate()
        .short_code("600496")
        .msisdn("254700000000")
        .amount(1000)
        .command_id(mpesa::CommandId::CustomerPayBillOnline) // optional, defaults to `CommandId::CustomerPayBillOnline`
        .bill_ref_number("Your_BillRefNumber") // optional, defaults to "None"
        .send()
        .await;

    assert!(response.is_ok())
}

pub fn account_balance<'a>( &'a self, initiator_name: &'a str, ) -> AccountBalanceBuilder<'a>

Account Balance

The Account Balance API is used to request the account balance of a short code. This can be used for both B2C, buy goods and pay bill accounts.

Requires an initiator_name. Returns an AccountBalanceBuilder for enquiring the balance on an MPESA Buy-Goods.

Safaricom API docs Account Balance reference

Example

use mpesa::{Environment, Mpesa};

#[tokio::main]
async fn main() {
    dotenvy::dotenv().ok();

    let client = Mpesa::new(
        dotenvy::var("CONSUMER_KEY").unwrap(),
        dotenvy::var("CONSUMER_SECRET").unwrap(),
        Environment::Sandbox,
    );

    let response = client
        .account_balance("testapi496")
        .result_url("https://testdomain.com/err")
        .timeout_url("https://testdomain.com/ok")
        .party_a("600496")
        .command_id(mpesa::CommandId::AccountBalance) // optional, defaults to `CommandId::AccountBalance`
        .identifier_type(mpesa::IdentifierTypes::ShortCode) // optional, defaults to `IdentifierTypes::ShortCode`
        .remarks("Your Remarks") // optional, defaults to "None"
        .send()
        .await;

    assert!(response.is_ok())
}

pub fn express_request(&self) -> MpesaExpressBuilder<'_>

Mpesa Express

Mpese Express Request

Lipa na M-PESA online API also known as M-PESA express (STK Push/NI push) is a Merchant/Business initiated C2B (Customer to Business) Payment.

Once you, our merchant integrate with the API, you will be able to send a payment prompt on the customer’s phone (Popularly known as STK Push Prompt) to your customer’s M-PESA registered phone number requesting them to enter their M-PESA pin to authorize and complete payment.

Requires a business_short_code - The organization shortcode used to receive the transaction and returns a MpesaExpressRequestBuilder struct

Safaricom API docs reference

Example

use mpesa::{Environment, Mpesa};

#[tokio::main]
async fn main() -> Result<(), Box<dyn std::error::Error>> {
    dotenvy::dotenv().ok();

    let client = Mpesa::new(
        dotenvy::var("CONSUMER_KEY").unwrap(),
        dotenvy::var("CONSUMER_SECRET").unwrap(),
        Environment::Sandbox,
    );

    let response = client
        .express_request()
        .business_short_code("174379")
        .phone_number("254708374149")
        .party_a("600584")
        .party_b("174379")
        .amount(500)
        .try_callback_url("https://test.example.com/api")?
        .account_ref("Test")
        .transaction_type(mpesa::CommandId::CustomerPayBillOnline) // Optional, defaults to `CommandId::CustomerPayBillOnline`
        .transaction_desc("Description") // Optional, defaults to "None"
        .build()?
        .send()
        .await;

    assert!(response.is_ok());

    Ok(())
}

Mpesa Express Query

M-PESA Express Query API checks the status of a Lipa Na M-PESA Online Payment.

Safaricom API docs reference

Query Example

use mpesa::{Environment, Mpesa};

#[tokio::main]
async fn main() -> Result<(), Box<dyn std::error::Error>> {
    dotenvy::dotenv().ok();

    let client = Mpesa::new(
        dotenvy::var("CLIENT_KEY").unwrap(),
        dotenvy::var("CLIENT_SECRET").unwrap(),
        Environment::Sandbox,
    );

    let response = client
        .express_query()
        .business_short_code("174379")
        .checkout_request_id("ws_CO_271120201234567891")
        .build()?
        .send()
        .await;

    assert!(response.is_ok());

    Ok(())
}

pub fn express_query(&self) -> MpesaExpressQueryBuilder<'_>

Mpesa Express

Mpese Express Request

Lipa na M-PESA online API also known as M-PESA express (STK Push/NI push) is a Merchant/Business initiated C2B (Customer to Business) Payment.

Once you, our merchant integrate with the API, you will be able to send a payment prompt on the customer’s phone (Popularly known as STK Push Prompt) to your customer’s M-PESA registered phone number requesting them to enter their M-PESA pin to authorize and complete payment.

Requires a business_short_code - The organization shortcode used to receive the transaction and returns a MpesaExpressRequestBuilder struct

Safaricom API docs reference

Example

use mpesa::{Environment, Mpesa};

#[tokio::main]
async fn main() -> Result<(), Box<dyn std::error::Error>> {
    dotenvy::dotenv().ok();

    let client = Mpesa::new(
        dotenvy::var("CONSUMER_KEY").unwrap(),
        dotenvy::var("CONSUMER_SECRET").unwrap(),
        Environment::Sandbox,
    );

    let response = client
        .express_request()
        .business_short_code("174379")
        .phone_number("254708374149")
        .party_a("600584")
        .party_b("174379")
        .amount(500)
        .try_callback_url("https://test.example.com/api")?
        .account_ref("Test")
        .transaction_type(mpesa::CommandId::CustomerPayBillOnline) // Optional, defaults to `CommandId::CustomerPayBillOnline`
        .transaction_desc("Description") // Optional, defaults to "None"
        .build()?
        .send()
        .await;

    assert!(response.is_ok());

    Ok(())
}

Mpesa Express Query

M-PESA Express Query API checks the status of a Lipa Na M-PESA Online Payment.

Safaricom API docs reference

Query Example

use mpesa::{Environment, Mpesa};

#[tokio::main]
async fn main() -> Result<(), Box<dyn std::error::Error>> {
    dotenvy::dotenv().ok();

    let client = Mpesa::new(
        dotenvy::var("CLIENT_KEY").unwrap(),
        dotenvy::var("CLIENT_SECRET").unwrap(),
        Environment::Sandbox,
    );

    let response = client
        .express_query()
        .business_short_code("174379")
        .checkout_request_id("ws_CO_271120201234567891")
        .build()?
        .send()
        .await;

    assert!(response.is_ok());

    Ok(())
}

pub fn transaction_reversal(&self) -> TransactionReversalBuilder<'_>

Transaction Reversal

Reverses a C2B M-Pesa transaction

Requires an initiator_name, the credential/ username used to authenticate the transaction request Returns a TransactionReversalBuilder

See more from the Safaricom API docs Reversals

Example

use mpesa::{Mpesa, Environment};

#[tokio::main]
async fn main() -> Result<(), Box<dyn std::error::Error>> {
    dotenvy::dotenv().ok();

    let client = Mpesa::new(
        dotenvy::var("CONSUMER_KEY").unwrap(),
        dotenvy::var("CONSUMER_SECRET").unwrap(),
        Environment::Sandbox,
    );

    let response = client
        .transaction_reversal()
        .initiator("testapi496")
        .try_result_url("https://testdomain.com/ok")?
        .try_timeout_url("https://testdomain.com/err")?
        .transaction_id("OEI2AK4Q16")
        .receiver_identifier_type(mpesa::IdentifierTypes::Reversal)
        .remarks("test")
        .amount(100)
        .receiver_party("600111")
        .build()?
        .send()
        .await;

    assert!(response.is_ok());

    Ok(())
}

pub fn transaction_status<'a>( &'a self, initiator_name: &'a str, ) -> TransactionStatusBuilder<'a>

Transaction Status

Queries the status of a B2B, B2C or C2B M-Pesa transaction.

Requires an initiator_name, the credential/ username used to authenticate the transaction request Returns a TransactionStatusBuilder.

See more from the Safaricom API docs Documentation

Example

use mpesa::{Environment, Mpesa};

#[tokio::main]
async fn main() {
    dotenvy::dotenv().ok();

    let client = Mpesa::new(
        dotenvy::var("CONSUMER_KEY").unwrap(),
        dotenvy::var("CONSUMER_SECRET").unwrap(),
        Environment::Sandbox,
    );

    let response = client
        .transaction_status("testapi496")
        .transaction_id("OEI2AK4Q16")
        .party_a("600496")
        .identifier_type(mpesa::IdentifierTypes::ShortCode) // optional, defaults to `IdentifierTypes::ShortCode`
        .remarks("Your Remarks") // optional, defaults to "None"
        .result_url("https://testdomain.com/err")
        .timeout_url("https://testdomain.com/ok")
        .send()
        .await;

    assert!(response.is_ok())
}

pub fn dynamic_qr(&self) -> DynamicQRBuilder<'_>

Dynamic QR

Generates a QR code that can be scanned by a M-Pesa customer to make payments.

Returns a DynamicQRBuilder

Safaricom API docs reference

Example

use mpesa::{Mpesa, Environment};

#[tokio::main]
async fn main() {
    dotenvy::dotenv().ok();

    let client = Mpesa::new(
        dotenvy::var("CONSUMER_KEY").unwrap(),
        dotenvy::var("CONSUMER_SECRET").unwrap(),
        Environment::Sandbox,
    );

    let response = client
        .dynamic_qr()
        .amount(2000)
        .credit_party_identifier("373132")
        .merchant_name("TEST SUPERMARKET")
        .ref_no("Invoice Test")
        .size("300")
        .transaction_type(mpesa::TransactionType::BG)
        .build()
        .unwrap()
        .send()
        .await;

    assert!(response.is_ok())
}